Move view and explorer functions to workspace
